refactor: move SchoolDashboard to its own file
Closes #137
